refactor: move CreateTransactionAnnotation to transaction.go :hammer:
