+++ components/ledger/internal/adapters/http/in/transaction.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/LerianStudio/midaz/v3/components/ledger/internal/services/command"
 	"github.com/LerianStudio/midaz/v3/components/ledger/internal/services/query"
+	"github.com/LerianStudio/midaz/v3/pkg/constant"
 	pkgTransaction "github.com/LerianStudio/midaz/v3/pkg/transaction"
 
 	"github.com/gofiber/fiber/v2"
@@ -50,9 +51,47 @@ func (handler *TransactionHandler) CreateTransactionJSON(p any, c *fiber.Ctx) er
 
 	input := p.(*pkgTransaction.CreateTransactionInput)
 	transactionInput := input.BuildTransaction()
-	logSafePayload(ctx, logger, "Request to create a transaction", transactionInput)
 
+	logSafePayload(ctx, logger, "Request to create a transaction", transactionInput)
 	recordSafePayloadAttributes(span, transactionInput)
 
 	return handler.createTransaction(c, *transactionInput, transactionInput.InitialStatus())
+}
+
+// CreateTransactionAnnotation method that create transaction using JSON
+//
+//	@Summary		Create a Transaction Annotation using JSON
+//	@Description	Create a Transaction Annotation with the input payload
+//	@Tags			Transactions
+//	@Accept			json
+//	@Produce		json
+//	@Param			Authorization	header		string						true	"Authorization Bearer Token"
+//	@Param			X-Request-Id	header		string						false	"Request ID"
+//	@Param			organization_id	path		string						true	"Organization ID"
+//	@Param			ledger_id		path		string						true	"Ledger ID"
+//	@Param			transaction		body		transaction.CreateTransactionInput	true	"Transaction Input"
+//	@Success		201				{object}	Transaction
+//	@Failure		400				{object}	mmodel.Error	"Invalid input, validation errors"
+//	@Failure		401				{object}	mmodel.Error	"Unauthorized access"
+//	@Failure		403				{object}	mmodel.Error	"Forbidden access"
+//	@Failure		422				{object}	mmodel.Error	"Unprocessable Entity, validation errors"
+//	@Failure		500				{object}	mmodel.Error	"Internal server error"
+//	@Router			/v1/organizations/{organization_id}/ledgers/{ledger_id}/transactions/annotation [post]
+func (handler *TransactionHandler) CreateTransactionAnnotation(p any, c *fiber.Ctx) error {
+	ctx := c.UserContext()
+
+	logger, tracer, _, _ := libCommons.NewTrackingFromContext(ctx)
+
+	ctx, span := tracer.Start(ctx, "handler.create_transaction_annotation")
+	defer span.End()
+
+	c.SetUserContext(ctx)
+
+	input := p.(*pkgTransaction.CreateTransactionInput)
+	transactionInput := input.BuildTransaction()
+
+	logSafePayload(ctx, logger, "Create a transaction annotation without an affected balance", transactionInput)
+	recordSafePayloadAttributes(span, transactionInput)
+
+	return handler.createTransaction(c, *transactionInput, constant.NOTED)
 }
