+++ components/ledger/internal/adapters/http/in/transaction_create_handlers.go
@@ -20,43 +20,6 @@ import (
 	"github.com/gofiber/fiber/v2"
 )
 
-// CreateTransactionAnnotation method that create transaction using JSON
-//
-//	@Summary		Create a Transaction Annotation using JSON
-//	@Description	Create a Transaction Annotation with the input payload
-//	@Tags			Transactions
-//	@Accept			json
-//	@Produce		json
-//	@Param			Authorization	header		string						true	"Authorization Bearer Token"
-//	@Param			X-Request-Id	header		string						false	"Request ID"
-//	@Param			organization_id	path		string						true	"Organization ID"
-//	@Param			ledger_id		path		string						true	"Ledger ID"
-//	@Param			transaction		body		transaction.CreateTransactionInput	true	"Transaction Input"
-//	@Success		201				{object}	Transaction
-//	@Failure		400				{object}	mmodel.Error	"Invalid input, validation errors"
-//	@Failure		401				{object}	mmodel.Error	"Unauthorized access"
-//	@Failure		403				{object}	mmodel.Error	"Forbidden access"
-//	@Failure		422				{object}	mmodel.Error	"Unprocessable Entity, validation errors"
-//	@Failure		500				{object}	mmodel.Error	"Internal server error"
-//	@Router			/v1/organizations/{organization_id}/ledgers/{ledger_id}/transactions/annotation [post]
-func (handler *TransactionHandler) CreateTransactionAnnotation(p any, c *fiber.Ctx) error {
-	ctx := c.UserContext()
-
-	logger, tracer, _, _ := libCommons.NewTrackingFromContext(ctx)
-
-	ctx, span := tracer.Start(ctx, "handler.create_transaction")
-	defer span.End()
-
-	c.SetUserContext(ctx)
-
-	input := p.(*pkgTransaction.CreateTransactionInput)
-	transactionInput := input.BuildTransaction()
-	logSafePayload(ctx, logger, "Create a transaction annotation without an affected balance", transactionInput)
-	recordSafePayloadAttributes(span, transactionInput)
-
-	return handler.createTransaction(c, *transactionInput, constant.NOTED)
-}
-
 // CreateTransactionInflow method that creates a transaction without specifying a source
 //
 //	@Summary		Create a Transaction without passing from source
